+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 most recent version is listed first.
 
 
+## **version:** v0.8.0
+- Add ability to use different codecs for different messages while using the same `naz` client: https://github.com/komuw/naz/pull/201   
+  With this change, message encoding is now a property of the message and not the client.
+
 
 ## **version:** v0.7.9
 - During shutdown, only close the writer without closing the reader end of the transport: https://github.com/komuw/naz/pull/198